renaming clientside names to match files
